requirements.txt
qrcode
requests
reportlab
spotipy
matplotlib
python-dotenv